Map.cs: thay doi mang thanh kieu so int
GameplayScreen.cs: sửa lỗi sai anh văn: You map HAS errors ( chứ ko phải have). Và đã đổi thành Your sudoku has error

MainMenuScreen.cs: thêm OptionsMenuScreen vào, thêm LevelScreen vào

OptionsMenuScreen.cs: đã thêm vào 3 MenuEntry: Sound, Music và Back. Hiện tại chưa xử lý dc vụ On và Off cho toàn bô chương trình vì còn phải đợi GameplayScreen làm phần âm thanh

+++ Game_Sudoku/Game_Sudoku/Game_Sudoku/Map/Map.cs
@@ -7,8 +7,9 @@ namespace Game_Sudoku.Map
 {
     class Map
     {
-        public float[,] MatrixMap = 
-            new float[9, 9] {
+
+        public int[,] MatrixMap = 
+            new int[9, 9] {
                             {4,0,0,8,0,7,0,0,2},
                             {0,0,2,4,0,5,9,0,0},
                             {0,0,8,0,6,0,7,0,0},
@@ -18,6 +19,7 @@ namespace Game_Sudoku.Map
                             {0,0,0,0,4,0,6,0,0},
                             {0,0,9,6,0,3,1,0,0},
                             {6,0,0,2,0,8,0,0,5}
+
                           };
 
     }
